hanyados double

+++ helloWorld/src/helloWorld/Main.java
@@ -24,7 +24,7 @@ public class Main {
 		int kivonas=a-b;
 		int osszead= a+b;
 		int kerulet= a+b+c;
-		double hanyados=a/b;
+		double hanyados=(double) a/b;
 		
 		System.out.println("Kerület: "+kerulet);
 		System.out.println("különbség: "+kivonas);
